requirements.txt
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
psycopg==3.1.8
psycopg-binary==3.1.8
pydantic==1.10.7
pytest==7.2.2
python-dotenv==1.0.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.8
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
pyston_lite_autoload